Move MemberIdBundle.sign() to SignatureBundle.sign()
