+++ src/integrationTests/main.test.ts
@@ -7,6 +7,7 @@ import {
   serialiseMemberIdBundle,
   MemberIdBundle,
   verify,
+  SignatureBundle,
 } from '../index.js';
 import { MEMBER_KEY_PAIR, MEMBER_NAME } from '../testUtils/veraStubs/member.js';
 import { arrayBufferFrom } from '../testUtils/buffers.js';
@@ -43,16 +44,18 @@ const MEMBER_ID_BUNDLE = MemberIdBundle.deserialise(MEMBER_ID_BUNDLE_SERIALISED)
 
 describe('main', () => {
   test('Valid signature bundle', async () => {
-    const signatureBundle = await MEMBER_ID_BUNDLE.sign(
+    const signatureBundle = await SignatureBundle.sign(
       PLAINTEXT,
       VERAID_OIDS.TEST_SERVICE,
+      MEMBER_ID_BUNDLE,
       MEMBER_KEY_PAIR.privateKey,
       EXPIRY_DATE,
     );
+    const signatureBundleSerialised = signatureBundle.serialise();
 
     const { plaintext, member } = await verify(
       PLAINTEXT,
-      signatureBundle,
+      signatureBundleSerialised,
       VERAID_OIDS.TEST_SERVICE,
     );
 
@@ -63,30 +66,34 @@ describe('main', () => {
 
   test('Invalid signature', async () => {
     const otherMemberKeyPair = await generateRsaKeyPair();
-    const signatureBundle = await MEMBER_ID_BUNDLE.sign(
+    const signatureBundle = await SignatureBundle.sign(
       PLAINTEXT,
       VERAID_OIDS.TEST_SERVICE,
+      MEMBER_ID_BUNDLE,
       otherMemberKeyPair.privateKey,
       EXPIRY_DATE,
     );
+    const signatureBundleSerialised = signatureBundle.serialise();
 
     await expect(async () =>
-      verify(PLAINTEXT, signatureBundle, VERAID_OIDS.TEST_SERVICE),
+      verify(PLAINTEXT, signatureBundleSerialised, VERAID_OIDS.TEST_SERVICE),
     ).rejects.toThrow(VeraidError);
   });
 
   test('Different service', async () => {
     const otherMemberKeyPair = await generateRsaKeyPair();
-    const signatureBundle = await MEMBER_ID_BUNDLE.sign(
+    const signatureBundle = await SignatureBundle.sign(
       PLAINTEXT,
       VERAID_OIDS.TEST_SERVICE,
+      MEMBER_ID_BUNDLE,
       otherMemberKeyPair.privateKey,
       EXPIRY_DATE,
     );
+    const signatureBundleSerialised = signatureBundle.serialise();
     const differentService = `${VERAID_OIDS.TEST_SERVICE}.42`;
 
-    await expect(async () => verify(PLAINTEXT, signatureBundle, differentService)).rejects.toThrow(
-      VeraidError,
-    );
+    await expect(async () =>
+      verify(PLAINTEXT, signatureBundleSerialised, differentService),
+    ).rejects.toThrow(VeraidError);
   });
 });
