+++ src/lib/memberIdBundle/MemberIdBundle.spec.ts
@@ -1,15 +1,7 @@
 import { jest } from '@jest/globals';
 import { AsnParser, AsnSerializer } from '@peculiar/asn1-schema';
 import { Certificate as CertificateSchema } from '@peculiar/asn1-x509';
-import { subSeconds, setMilliseconds } from 'date-fns';
-import type { OctetString } from 'asn1js';
-import {
-  Attribute,
-  type ContentInfo,
-  SignedData as SignedDataSchema,
-  EncapsulatedContent,
-  type SignerInfo,
-} from '@peculiar/asn1-cms';
+import { subSeconds } from 'date-fns';
 
 import { generateMemberIdFixture } from '../../testUtils/veraStubs/memberIdFixture.js';
 import {
@@ -34,11 +26,6 @@ import { MemberIdBundleSchema } from '../schemas/MemberIdBundleSchema.js';
 import Certificate from '../utils/x509/Certificate.js';
 import { issueMemberCertificate } from '../pki/member.js';
 import { MEMBER_KEY_PAIR, MEMBER_NAME } from '../../testUtils/veraStubs/member.js';
-import { SignatureBundleSchema } from '../schemas/SignatureBundleSchema.js';
-import { SignedData } from '../utils/cms/SignedData.js';
-import { CMS_OIDS, VERAID_OIDS } from '../oids.js';
-import { SignatureMetadataSchema } from '../schemas/SignatureMetadataSchema.js';
-import { derDeserialize } from '../utils/asn1.js';
 
 import { MemberIdBundle } from './MemberIdBundle.js';
 import { serialiseMemberIdBundle } from './serialisation.js';
@@ -50,9 +37,6 @@ const dnssecChain = new DnssecChainSchema(
 );
 const orgCertificateSchema = AsnParser.parse(orgCertificateSerialised, CertificateSchema);
 const memberCertificateSchema = AsnParser.parse(memberCertificateSerialised, CertificateSchema);
-
-// Create a plaintext for testing
-const PLAINTEXT = arrayBufferFrom('Hello world');
 
 describe('MemberIdBundle', () => {
   describe('serialise', () => {
@@ -371,274 +355,6 @@ describe('MemberIdBundle', () => {
         );
 
         expect(user).toBeUndefined();
-      });
-    });
-  });
-
-  describe('sign', () => {
-    test('Version should be 0', async () => {
-      const bundle = new MemberIdBundle(dnssecChain, orgCertificateSchema, memberCertificateSchema);
-
-      const signatureSerialised = await bundle.sign(
-        PLAINTEXT,
-        SERVICE_OID,
-        MEMBER_KEY_PAIR.privateKey,
-        datePeriod.end,
-      );
-
-      const { version } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-      expect(version).toBe(0);
-    });
-
-    test('DNSSEC chain should be attached', async () => {
-      const bundle = new MemberIdBundle(dnssecChain, orgCertificateSchema, memberCertificateSchema);
-
-      const signatureSerialised = await bundle.sign(
-        PLAINTEXT,
-        SERVICE_OID,
-        MEMBER_KEY_PAIR.privateKey,
-        datePeriod.end,
-      );
-
-      const { dnssecChain: signatureDnssecChain } = AsnParser.parse(
-        signatureSerialised,
-        SignatureBundleSchema,
-      );
-      expect(Buffer.from(AsnSerializer.serialize(signatureDnssecChain))).toStrictEqual(
-        Buffer.from(AsnSerializer.serialize(dnssecChain)),
-      );
-    });
-
-    test('Organisation certificate should be attached', async () => {
-      const bundle = new MemberIdBundle(dnssecChain, orgCertificateSchema, memberCertificateSchema);
-
-      const signatureSerialised = await bundle.sign(
-        PLAINTEXT,
-        SERVICE_OID,
-        MEMBER_KEY_PAIR.privateKey,
-        datePeriod.end,
-      );
-
-      const { organisationCertificate } = AsnParser.parse(
-        signatureSerialised,
-        SignatureBundleSchema,
-      );
-      expect(Buffer.from(AsnSerializer.serialize(organisationCertificate))).toStrictEqual(
-        Buffer.from(orgCertificateSerialised),
-      );
-    });
-
-    describe('Signature', () => {
-      function getSignedData(contentInfo: ContentInfo) {
-        expect(contentInfo.contentType).toStrictEqual(CMS_OIDS.SIGNED_DATA);
-        return AsnParser.parse(contentInfo.content, SignedDataSchema);
-      }
-
-      test('Plaintext should be signed with specified private key', async () => {
-        const bundle = new MemberIdBundle(
-          dnssecChain,
-          orgCertificateSchema,
-          memberCertificateSchema,
-        );
-
-        const signatureSerialised = await bundle.sign(
-          PLAINTEXT,
-          SERVICE_OID,
-          MEMBER_KEY_PAIR.privateKey,
-          datePeriod.end,
-        );
-
-        const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-        const signedData = SignedData.deserialize(AsnSerializer.serialize(signature));
-        await signedData.verify(PLAINTEXT);
-        const memberCertificate = Certificate.deserialize(memberCertificateSerialised);
-        expect(signedData.signerCertificate!.isEqual(memberCertificate)).toBeTrue();
-      });
-
-      test('Member certificate should be attached', async () => {
-        const bundle = new MemberIdBundle(
-          dnssecChain,
-          orgCertificateSchema,
-          memberCertificateSchema,
-        );
-
-        const signatureSerialised = await bundle.sign(
-          PLAINTEXT,
-          SERVICE_OID,
-          MEMBER_KEY_PAIR.privateKey,
-          datePeriod.end,
-        );
-
-        const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-        const { certificates } = getSignedData(signature);
-        const attachedCertsSerialised = certificates!.map((cert) =>
-          Buffer.from(AsnSerializer.serialize(cert)),
-        );
-        expect(attachedCertsSerialised).toContainEqual(Buffer.from(memberCertificateSerialised));
-      });
-
-      test('Plaintext should be detached by default', async () => {
-        const bundle = new MemberIdBundle(
-          dnssecChain,
-          orgCertificateSchema,
-          memberCertificateSchema,
-        );
-
-        const signatureSerialised = await bundle.sign(
-          PLAINTEXT,
-          SERVICE_OID,
-          MEMBER_KEY_PAIR.privateKey,
-          datePeriod.end,
-        );
-
-        const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-        const { encapContentInfo } = getSignedData(signature);
-        expect(encapContentInfo.eContent).toBeUndefined();
-      });
-
-      test('Plaintext should be attached if requested', async () => {
-        const bundle = new MemberIdBundle(
-          dnssecChain,
-          orgCertificateSchema,
-          memberCertificateSchema,
-        );
-
-        const signatureSerialised = await bundle.sign(
-          PLAINTEXT,
-          SERVICE_OID,
-          MEMBER_KEY_PAIR.privateKey,
-          datePeriod.end,
-          { shouldEncapsulatePlaintext: true },
-        );
-
-        const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-        const { encapContentInfo } = getSignedData(signature);
-        expect(encapContentInfo.eContent).toBeInstanceOf(EncapsulatedContent);
-        const encapsulatedContentSerialised = encapContentInfo.eContent!.any!;
-        const encapsulatedContentAsn1 = derDeserialize(
-          encapsulatedContentSerialised,
-        ) as OctetString;
-        expect(new Uint8Array(encapsulatedContentAsn1.getValue())).toStrictEqual(
-          new Uint8Array(PLAINTEXT),
-        );
-      });
-
-      describe('Metadata', () => {
-        function getSignedAttribute(signerInfo: SignerInfo, attributeOid: string) {
-          const matchingAttribute = signerInfo.signedAttrs!.find(
-            (attribute) => attribute.attrType === attributeOid,
-          );
-          expect(matchingAttribute).toBeInstanceOf(Attribute);
-          return matchingAttribute!.attrValues;
-        }
-
-        function getSignatureMetadata(signature: ContentInfo) {
-          const { signerInfos } = getSignedData(signature);
-          const attributeValues = getSignedAttribute(
-            signerInfos[0],
-            VERAID_OIDS.SIGNATURE_METADATA_ATTR,
-          );
-          expect(attributeValues).toHaveLength(1);
-          return AsnParser.parse(attributeValues[0], SignatureMetadataSchema);
-        }
-
-        test('Service OID should be attached', async () => {
-          const bundle = new MemberIdBundle(
-            dnssecChain,
-            orgCertificateSchema,
-            memberCertificateSchema,
-          );
-
-          const signatureSerialised = await bundle.sign(
-            PLAINTEXT,
-            SERVICE_OID,
-            MEMBER_KEY_PAIR.privateKey,
-            datePeriod.end,
-          );
-
-          const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-          const { serviceOid } = getSignatureMetadata(signature);
-          expect(serviceOid).toStrictEqual(SERVICE_OID);
-        });
-
-        test('Expiry date should be attached', async () => {
-          const bundle = new MemberIdBundle(
-            dnssecChain,
-            orgCertificateSchema,
-            memberCertificateSchema,
-          );
-
-          const signatureSerialised = await bundle.sign(
-            PLAINTEXT,
-            SERVICE_OID,
-            MEMBER_KEY_PAIR.privateKey,
-            datePeriod.end,
-          );
-
-          const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-          const { validityPeriod } = getSignatureMetadata(signature);
-          expect(validityPeriod.end).toStrictEqual(datePeriod.end);
-        });
-
-        test('Start date should default to the current time', async () => {
-          const beforeSignatureDate = setMilliseconds(new Date(), 0);
-          const bundle = new MemberIdBundle(
-            dnssecChain,
-            orgCertificateSchema,
-            memberCertificateSchema,
-          );
-
-          const signatureSerialised = await bundle.sign(
-            PLAINTEXT,
-            SERVICE_OID,
-            MEMBER_KEY_PAIR.privateKey,
-            datePeriod.end,
-          );
-          const afterSignatureDate = new Date();
-
-          const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-          const { validityPeriod } = getSignatureMetadata(signature);
-          expect(validityPeriod.start).toBeBetween(beforeSignatureDate, afterSignatureDate);
-        });
-
-        test('Any explicit start date should be honoured', async () => {
-          const startDate = subSeconds(datePeriod.start, 1);
-          const bundle = new MemberIdBundle(
-            dnssecChain,
-            orgCertificateSchema,
-            memberCertificateSchema,
-          );
-
-          const signatureSerialised = await bundle.sign(
-            PLAINTEXT,
-            SERVICE_OID,
-            MEMBER_KEY_PAIR.privateKey,
-            datePeriod.end,
-            { startDate },
-          );
-
-          const { signature } = AsnParser.parse(signatureSerialised, SignatureBundleSchema);
-          const { validityPeriod } = getSignatureMetadata(signature);
-          expect(validityPeriod.start).toStrictEqual(startDate);
-        });
-
-        test('Start date after expiry date should be refused', async () => {
-          const invalidExpiryDate = subSeconds(datePeriod.start, 1);
-          const bundle = new MemberIdBundle(
-            dnssecChain,
-            orgCertificateSchema,
-            memberCertificateSchema,
-          );
-
-          await expect(async () =>
-            bundle.sign(PLAINTEXT, SERVICE_OID, MEMBER_KEY_PAIR.privateKey, invalidExpiryDate, {
-              startDate: datePeriod.start,
-            }),
-          ).rejects.toThrowWithMessage(
-            VeraidError,
-            'Signature start date cannot be after expiry date',
-          );
-        });
       });
     });
   });
