+++ src/lib/memberIdBundle/MemberIdBundle.ts
@@ -1,9 +1,6 @@
 import type { TrustAnchor } from '@relaycorp/dnssec';
 import type { Certificate as CertificateSchema } from '@peculiar/asn1-x509';
 import { AsnParser, AsnSerializer } from '@peculiar/asn1-schema';
-import { ContentInfo } from '@peculiar/asn1-cms';
-import { Attribute } from 'pkijs';
-import type { Sequence } from 'asn1js';
 
 import type { DnssecChainSchema } from '../schemas/DnssecChainSchema.js';
 import type { DatePeriod } from '../dates.js';
@@ -15,55 +12,6 @@ import { MemberIdBundleSchema } from '../schemas/MemberIdBundleSchema.js';
 import type { Member } from '../Member.js';
 import { BOT_NAME } from '../pki/member.js';
 import { validateUserName } from '../idValidation.js';
-import { SignatureBundleSchema } from '../schemas/SignatureBundleSchema.js';
-import { VERAID_OIDS } from '../oids.js';
-import { SignatureMetadataSchema } from '../schemas/SignatureMetadataSchema.js';
-import { DatePeriodSchema } from '../schemas/DatePeriodSchema.js';
-import { derDeserialize } from '../utils/asn1.js';
-import { SignedData } from '../utils/cms/SignedData.js';
-import type { SignatureOptions } from '../SignatureOptions.js';
-
-function generateMetadata(serviceOid: string, startDate: Date, expiryDate: Date): Sequence {
-  if (expiryDate < startDate) {
-    throw new VeraidError('Signature start date cannot be after expiry date');
-  }
-
-  const metadataSchema = new SignatureMetadataSchema();
-
-  metadataSchema.serviceOid = serviceOid;
-
-  const datePeriod = new DatePeriodSchema();
-  datePeriod.start = startDate;
-  datePeriod.end = expiryDate;
-  metadataSchema.validityPeriod = datePeriod;
-
-  const serialisation = AsnSerializer.serialize(metadataSchema);
-  return derDeserialize(serialisation) as Sequence;
-}
-
-async function generateSignedData(
-  plaintext: ArrayBuffer,
-  memberCertificateSchema: CertificateSchema,
-  signingKey: CryptoKey,
-  serviceOid: string,
-  shouldEncapsulatePlaintext: boolean,
-  expiryDate: Date,
-  startDate?: Date,
-) {
-  const memberCertificate = Certificate.deserialize(
-    AsnSerializer.serialize(memberCertificateSchema),
-  );
-  const metadataSchema = generateMetadata(serviceOid, startDate ?? new Date(), expiryDate);
-  const metadataAttribute = new Attribute({
-    type: VERAID_OIDS.SIGNATURE_METADATA_ATTR,
-    values: [metadataSchema],
-  });
-  const signedData = await SignedData.sign(plaintext, signingKey, memberCertificate, [], {
-    extraSignedAttrs: [metadataAttribute],
-    encapsulatePlaintext: shouldEncapsulatePlaintext,
-  });
-  return AsnParser.parse(signedData.serialize(), ContentInfo);
-}
 
 async function verifyCertificateChain(
   orgCertificate: Certificate,
@@ -132,42 +80,6 @@ export class MemberIdBundle {
     bundle.organisationCertificate = this.orgCertificateSchema;
     bundle.dnssecChain = this.dnssecChainSchema;
     return AsnSerializer.serialize(bundle);
-  }
-
-  /**
-   * Create a signature for the specified plaintext using this member ID
-   * @param plaintext - The data to sign
-   * @param serviceOid - The OID of the service for which the signature is created
-   * @param signingKey - The private key corresponding to the member certificate
-   * @param expiryDate - The date when the signature expires
-   * @param options - Additional options for signature creation
-   * @param options.startDate - The date when the signature becomes valid (defaults to now)
-   * @param options.shouldEncapsulatePlaintext - Whether to include the plaintext in the signature (defaults to false)
-   * @returns A serialised SignatureBundle
-   */
-  public async sign(
-    plaintext: ArrayBuffer,
-    serviceOid: string,
-    signingKey: CryptoKey,
-    expiryDate: Date,
-    { startDate, shouldEncapsulatePlaintext }: Partial<SignatureOptions> = {},
-  ): Promise<ArrayBuffer> {
-    const signedDataSchema = await generateSignedData(
-      plaintext,
-      this.memberCertificateSchema,
-      signingKey,
-      serviceOid,
-      shouldEncapsulatePlaintext ?? false,
-      expiryDate,
-      startDate,
-    );
-
-    const signatureSchema = new SignatureBundleSchema();
-    signatureSchema.version = 0;
-    signatureSchema.dnssecChain = this.dnssecChainSchema;
-    signatureSchema.organisationCertificate = this.orgCertificateSchema;
-    signatureSchema.signature = signedDataSchema;
-    return AsnSerializer.serialize(signatureSchema);
   }
 
   /**
